# contracts/product-api/v2/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "JsonValue": {},
...
      "SearchErrorEnvelopeDto": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/SearchErrorDto"
          }
        },
        "required": [
          "error"
        ],
        "title": "SearchErrorEnvelopeDto",
        "type": "object"
      },
      "SearchExperimentResponseDto": {
        "additionalProperties": false,
        "properties": {
          "experiment": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "title": "Experiment",
            "type": "object"
          },
          "experiment_fingerprint": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Experiment Fingerprint",
            "type": "string",
            "x-onlyalpha-reference-kind": "SEARCH_EXPERIMENT",
            "x-onlyalpha-reference-locator-kind": "SHA256",
            "x-onlyalpha-reference-schema-version": 1
          },
          "method": {
            "enum": [
              "SYMBOLIC",
              "PARAMETER"
            ],
            "title": "Method",
            "type": "string"
          },
          "schema_version": {
            "const": 1,
            "default": 1,
            "title": "Schema Version",
            "type": "integer"
          }
        },
        "required": [
          "experiment_fingerprint",
          "method",
          "experiment"
        ],
        "title": "SearchExperimentResponseDto",
        "type": "object"
      },
      "SearchLedgerResponseDto": {
        "additionalProperties": false,
        "properties": {
          "experiment_fingerprint": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Experiment Fingerprint",
            "type": "string",
            "x-onlyalpha-reference-kind": "SEARCH_EXPERIMENT",
            "x-onlyalpha-reference-locator-kind": "SHA256",
            "x-onlyalpha-reference-schema-version": 1
          },
          "ledger": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "title": "Ledger",
            "type": "object"
          },
          "method": {
            "enum": [
              "SYMBOLIC",
              "PARAMETER"
            ],
            "title": "Method",
            "type": "string"
          },
          "schema_version": {
            "const": 1,
            "default": 1,
            "title": "Schema Version",
            "type": "integer"
          }
        },
        "required": [
          "experiment_fingerprint",
          "method",
          "ledger"
        ],
        "title": "SearchLedgerResponseDto",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "OnlyAlpha Product API",
    "version": "2"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v2/research/search/experiments/{experiment_fingerprint}": {
      "get": {
        "operationId": "get_search_experiment_v2",
        "parameters": [
          {
            "in": "path",
            "name": "experiment_fingerprint",
            "required": true,
            "schema": {
              "pattern": "^[0-9a-f]{64}$",
              "title": "Experiment Fingerprint",
              "type": "string",
              "x-onlyalpha-reference-kind": "SEARCH_EXPERIMENT",
              "x-onlyalpha-reference-locator-kind": "SHA256",
              "x-onlyalpha-reference-schema-version": 1
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchExperimentResponseDto"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchErrorEnvelopeDto"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchErrorEnvelopeDto"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchErrorEnvelopeDto"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchErrorEnvelopeDto"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchErrorEnvelopeDto"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Get Experiment",
        "tags": [
          "search-product"
        ],
        "x-onlyalpha-agent-operation": {
          "identity_requirements": [
            "experiment_fingerprint"
          ],
          "owning_authority_references": [
            {
              "locator_kind": "SHA256",
              "reference_kind": "SEARCH_EXPERIMENT",
              "reference_schema_version": 1,
              "response_field": "experiment_fingerprint"
            }
          ],
          "recovery_class": "MUTABLE_OBSERVATION_QUERY",
          "requires_product_command_id": false,
          "schema_version": 1,
          "tool_class": "SEARCH_QUERY"
        }
      }
    },
    "/api/v2/research/search/experiments/{experiment_fingerprint}/ledger": {
      "get": {
        "operationId": "get_search_iteration_ledger_v2",
        "parameters": [
          {
            "in": "path",
            "name": "experiment_fingerprint",
            "required": true,
            "schema": {
              "pattern": "^[0-9a-f]{64}$",
              "title": "Experiment Fingerprint",
              "type": "string",
              "x-onlyalpha-reference-kind": "SEARCH_EXPERIMENT",
              "x-onlyalpha-reference-locator-kind": "SHA256",
              "x-onlyalpha-reference-schema-version": 1
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchLedgerResponseDto"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchErrorEnvelopeDto"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchErrorEnvelopeDto"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchErrorEnvelopeDto"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchErrorEnvelopeDto"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchErrorEnvelopeDto"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Get Ledger",
        "tags": [
          "search-product"
        ],
        "x-onlyalpha-agent-operation": {
          "identity_requirements": [
            "experiment_fingerprint"
          ],
          "owning_authority_references": [
            {
              "locator_kind": "SHA256",
              "reference_kind": "SEARCH_EXPERIMENT",
              "reference_schema_version": 1,
              "response_field": "experiment_fingerprint"
            }
          ],
          "recovery_class": "MUTABLE_OBSERVATION_QUERY",
          "requires_product_command_id": false,
          "schema_version": 1,
          "tool_class": "SEARCH_QUERY"
        }
      }
    },
    "/api/v2/research/search/experiments/{experiment_fingerprint}/terminal": {
      "get": {
        "operationId": "get_search_terminal_v2",
        "parameters": [
          {
            "in": "path",
            "name": "experiment_fingerprint",
            "required": true,
            "schema": {
              "pattern": "^[0-9a-f]{64}$",
              "title": "Experiment Fingerprint",
              "type": "string",
              "x-onlyalpha-reference-kind": "SEARCH_EXPERIMENT",
              "x-onlyalpha-reference-locator-kind": "SHA256",
              "x-onlyalpha-reference-schema-version": 1
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchTerminalResponseDto"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchErrorEnvelopeDto"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchErrorEnvelopeDto"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchErrorEnvelopeDto"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchErrorEnvelopeDto"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchErrorEnvelopeDto"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Get Terminal",
        "tags": [
          "search-product"
        ],
        "x-onlyalpha-agent-operation": {
          "identity_requirements": [
            "experiment_fingerprint"
          ],
          "owning_authority_references": [
            {
              "locator_kind": "SHA256",
              "reference_kind": "SEARCH_EXPERIMENT",
              "reference_schema_version": 1,
              "response_field": "experiment_fingerprint"
            }
          ],
          "recovery_class": "MUTABLE_OBSERVATION_QUERY",
          "requires_product_command_id": false,
          "schema_version": 1,
          "tool_class": "SEARCH_QUERY"
        }
      }
    },
    "/api/v2/research/search/parameter-experiments": {
      "post": {
        "operationId": "submit_parameter_search_experiment_v2",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ParameterSearchSubmitRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchCommandResponseDto"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "Idempotency-Key": {
                "description": "Exact admitted Product Command identity",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchErrorEnvelopeDto"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchErrorEnvelopeDto"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchErrorEnvelopeDto"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchErrorEnvelopeDto"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchErrorEnvelopeDto"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Submit Parameter",
        "tags": [
          "search-product"
        ],
        "x-onlyalpha-agent-operation": {
          "identity_requirements": [
            "catalog_generation_fingerprint",
            "dataset_snapshot_fingerprint",
            "runtime_generation_fingerprint"
          ],
          "owning_authority_references": [
            {
              "locator_kind": "SHA256",
              "reference_kind": "SEARCH_EXPERIMENT",
              "reference_schema_version": 1,
              "response_field": "experiment_fingerprint"
            }
          ],
          "product_command_id_transport": {
            "in": "header",
            "name": "Idempotency-Key"
          },
          "recovery_class": "IDEMPOTENT_COMMAND",
          "requires_product_command_id": true,
          "schema_version": 1,
          "tool_class": "PARAMETER_SEARCH"
        }
      }
    },
    "/api/v2/research/search/parameter-experiments/advance": {
      "post": {
        "operationId": "advance_parameter_search_experiment_v2",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchAdvanceRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchCommandResponseDto"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "Idempotency-Key": {
                "description": "Exact admitted Product Command identity",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchErrorEnvelopeDto"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchErrorEnvelopeDto"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchErrorEnvelopeDto"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchErrorEnvelopeDto"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchErrorEnvelopeDto"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Advance Parameter",
        "tags": [
          "search-product"
        ],
        "x-onlyalpha-agent-operation": {
          "identity_requirements": [
            "experiment_fingerprint"
          ],
          "owning_authority_references": [
            {
              "locator_kind": "SHA256",
              "reference_kind": "SEARCH_EXPERIMENT",
              "reference_schema_version": 1,
              "response_field": "experiment_fingerprint"
            }
          ],
          "product_command_id_transport": {
            "in": "header",
            "name": "Idempotency-Key"
          },
          "recovery_class": "IDEMPOTENT_COMMAND",
          "requires_product_command_id": true,
          "schema_version": 1,
          "tool_class": "PARAMETER_SEARCH"
        }
      }
    },
    "/api/v2/research/search/symbolic-experiments": {
      "post": {
        "operationId": "submit_symbolic_search_experiment_v2",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SymbolicSearchSubmitRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchCommandResponseDto"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "Idempotency-Key": {
                "description": "Exact admitted Product Command identity",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchErrorEnvelopeDto"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchErrorEnvelopeDto"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchErrorEnvelopeDto"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchErrorEnvelopeDto"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchErrorEnvelopeDto"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Submit Symbolic",
        "tags": [
          "search-product"
        ],
        "x-onlyalpha-agent-operation": {
          "identity_requirements": [
            "catalog_generation_fingerprint",
            "dataset_snapshot_fingerprint",
            "runtime_generation_fingerprint"
          ],
          "owning_authority_references": [
            {
              "locator_kind": "SHA256",
              "reference_kind": "SEARCH_EXPERIMENT",
              "reference_schema_version": 1,
              "response_field": "experiment_fingerprint"
            }
          ],
          "product_command_id_transport": {
            "in": "header",
            "name": "Idempotency-Key"
          },
          "recovery_class": "IDEMPOTENT_COMMAND",
          "requires_product_command_id": true,
          "schema_version": 1,
          "tool_class": "SYMBOLIC_SEARCH"
        }
      }
    },
    "/api/v2/research/search/symbolic-experiments/advance": {
      "post": {
        "operationId": "advance_symbolic_search_experiment_v2",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchAdvanceRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchCommandResponseDto"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "Idempotency-Key": {
                "description": "Exact admitted Product Command identity",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchErrorEnvelopeDto"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchErrorEnvelopeDto"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchErrorEnvelopeDto"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchErrorEnvelopeDto"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchErrorEnvelopeDto"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Advance Symbolic",
        "tags": [
          "search-product"
        ],
        "x-onlyalpha-agent-operation": {
          "identity_requirements": [
            "experiment_fingerprint"
          ],
          "owning_authority_references": [
            {
              "locator_kind": "SHA256",
              "reference_kind": "SEARCH_EXPERIMENT",
              "reference_schema_version": 1,
              "response_field": "experiment_fingerprint"
            }
          ],
          "product_command_id_transport": {
            "in": "header",
            "name": "Idempotency-Key"
          },
          "recovery_class": "IDEMPOTENT_COMMAND",
          "requires_product_command_id": true,
          "schema_version": 1,
          "tool_class": "SYMBOLIC_SEARCH"
        }
      }
    },
...
  }
}